fix(transport): DEVICE_TYPE circular dependencies
