+++ packages/transport-bluetooth/src/client/bluetooth-api.ts
@@ -1,9 +1,9 @@
 import {
     AbstractApi,
     AbstractApiConstructorParams,
-    DEVICE_TYPE,
     OpenDeviceChannel,
 } from '@trezor/transport/src/api/abstract';
+import { DEVICE_TYPE } from '@trezor/transport/src/constants';
 import * as ERRORS from '@trezor/transport/src/errors';
 import { PathInternal } from '@trezor/transport/src/types';
 import { readMessageBuffer } from '@trezor/transport/src/utils/readMessageBuffer';
