+++ packages/transport-test/e2e/bridge/expect.ts
@@ -1,4 +1,4 @@
-import { DEVICE_TYPE } from '@trezor/transport/src/api/abstract';
+import { DEVICE_TYPE } from '@trezor/transport/src/constants';
 
 import { env } from './controller';
