+++ packages/transport/src/api/abstract.ts
@@ -16,17 +16,6 @@ export interface AbstractApiConstructorParams {
     logger?: Logger;
 }
 
-// https://github.dev/trezor/trezord-go/blob/db03d99230f5b609a354e3586f1dfc0ad6da16f7/core/core.go#L46-L47
-export enum DEVICE_TYPE {
-    TypeT1Hid = 0,
-    TypeT1Webusb = 1,
-    TypeT1WebusbBoot = 2,
-    TypeT2 = 3,
-    TypeT2Boot = 4,
-    TypeEmulator = 5,
-    TypeBluetooth = 6,
-}
-
 export type OpenDeviceChannel = 'read' | 'push-notification' | 'battery-level';
 
 type AccessLock = {
